{{cookiecutter.project_name}}/requirements.txt
absl-py==0.6.1
aniso8601==4.1.0
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.5
djangorestframework==3.9.0
drf-yasg==1.12.1
gast==0.2.2
grpcio==1.17.1
gunicorn==19.9.0
h5py==2.9.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
numpy==1.15.4
pandas==0.23.4
protobuf==3.6.1
psycopg2==2.7.6.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.21.0
ruamel.yaml==0.15.85
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
whitenoise==4.1.2
